# public/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "description",
        "version": "0.5.0",
        "title": "China Cipher Messenger Server",
        "contact": {
            "email": "t105590035@ntut.org.tw"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/ChinaCipher/messenger-server/blob/master/LICENSE"
        }
    },
    "host": "ccm.ntut.com.tw",
    "basePath": "/api",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/chat": {
            "get": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "取得聊天室列表",
                "description": "",
                "operationId": "get_rooms",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Chat.Rooms"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "建立聊天室",
                "description": "",
                "operationId": "create_room",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "所需的資料",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.Username"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Chat.Room"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            }
        },
        "/chat/{username}": {
            "get": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "取得聊天室資訊",
                "description": "",
                "operationId": "get_room",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Chat.Room"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "移除聊天室",
                "description": "",
                "operationId": "delete_room",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            }
        },
        "/chat/{username}/message": {
            "get": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "取得訊息列表",
                "description": "",
                "operationId": "get_messages",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "index",
                        "in": "query",
                        "description": "從第幾筆資料開始",
                        "type": "integer"
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "取幾筆資料",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Chat.Messages"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "發送訊息",
                "description": "",
                "operationId": "send_message",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "要發送的訊息",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Chat.Content"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Chat.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            }
        },
        "/chat/{username}/message/{messageId}": {
            "get": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "取得訊息",
                "description": "",
                "operationId": "get_message",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "訊息 ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Chat.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "修改訊息",
                "description": "",
                "operationId": "put_message",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "訊息 ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "所需的資料",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Chat.Content.Editable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Chat.Message"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ChatApi"
                ],
                "summary": "刪除訊息",
                "description": "",
                "operationId": "delete_message",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "使用者名稱",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "訊息 ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Test.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "訊息內容",
                    "type": "string"
                }
            }
        },
...
        "User.Profile": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "大頭貼網址",
                    "type": "string"
                },
                "username": {
                    "description": "使用者名稱",
                    "type": "string"
                },
                "nickname": {
                    "description": "使用者暱稱",
                    "type": "string"
                }
            }
        },
        "User.Username": {
            "type": "object",
            "properties": {
                "username": {
                    "description": "使用者名稱",
                    "type": "string"
                }
            }
        },
...
        "Chat.Rooms": {
            "type": "object",
            "properties": {
                "rooms": {
                    "description": "房間列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User.Profile"
                    }
                }
            }
        },
        "Chat.Room": {
            "type": "object",
            "properties": {
                "messageKey": {
                    "description": "加密過後的訊息金鑰",
                    "type": "string"
                }
            }
        },
        "Chat.Messages": {
            "type": "object",
            "properties": {
                "messages": {
                    "description": "訊息列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Chat.Message"
                    }
                }
            }
        },
        "Chat.Message": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "訊息 ID",
                    "type": "integer"
                },
                "sender": {
                    "description": "發送者名稱",
                    "type": "string"
                },
                "type": {
                    "description": "訊息種類",
                    "type": "string"
                },
                "content": {
                    "description": "訊息內容",
                    "type": "string"
                },
                "options": {
                    "description": "選填內容",
                    "type": "object"
                },
                "timestamp": {
                    "description": "時間戳記",
                    "type": "string"
                }
            }
        },
        "Chat.Content": {
            "type": "object",
            "properties": {
                "type": {
                    "description": "訊息種類",
                    "type": "string"
                },
                "content": {
                    "description": "訊息內容",
                    "type": "string"
                },
                "options": {
                    "description": "選填內容",
                    "type": "object"
                }
            }
        },
        "Chat.Content.Editable": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "訊息內容",
                    "type": "string"
                },
                "options": {
                    "description": "選填內容",
                    "type": "object"
                }
            }
        }
    }
}